backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.9
starlette>=0.46.0
anyio>=4.8.0

torch>=2.1.0
Pillow>=10.2.0
ffmpeg-python==0.2.0
openai>=1.0.0
python-dotenv==1.0.0
httpx>=0.28.0
groq>=0.4.0
google-generativeai>=0.8.0

# Optional heavy deps (comment out to speed up install)
# insightface>=0.7.3
# onnxruntime>=1.17.0
# ultralytics==8.3.0
# pyannote.audio==3.1.1
opencv-python-headless>=4.10.0

# For GLM-4.6V OCR (optional, large model)
transformers>=4.40.0
accelerate>=0.27.0

# Music & Audio Processing
ytmusicapi>=1.11.0
yt-dlp>=2023.10.0
librosa>=0.10.0
numpy>=1.24.0
scipy>=1.10.0

# TikTok Trends & Virality
TikTokApi>=6.1.0
playwright>=1.40.0
moviepy>=1.0.3
sentence-transformers>=2.2.2
